requirements.txt
streamlit
langchain
langchain-openai
langchain-core
python-dotenv
openai